clean up articleVCAccessesSchemeHandler with async + Swift Testing
