+++ WikipediaUnitTests/Code/ArticleViewControllerTests.swift
@@ -1,71 +1,47 @@
+import Foundation
+import Testing
+import UIKit
 @testable import Wikipedia
-import XCTest
 
-private class MockSchemeHandler: SchemeHandler {
-    
-    var accessed = false
-    
-    required init(scheme: String, session: Session) {
-        super.init(scheme: scheme, session: session)
-        let didReceiveDataCallback: ((WKURLSchemeTask, Data) -> Void)? = { [weak self] _, _ in
-            self?.accessed = true
+struct ArticleViewControllerTests {
+
+    @MainActor
+    @Test
+    func articleVCAccessesSchemeHandler() async throws {
+        await withCheckedContinuation { (continuation: CheckedContinuation<Void, Never>) in
+            ArticleTestHelpers.setupWithNetworkFixtures {
+                continuation.resume()
+            }
         }
-        self.didReceiveDataCallback = didReceiveDataCallback
-    }
-}
-
-class ArticleViewControllerTests: XCTestCase {
-    let timeout: TimeInterval = 10
-    
-    override func setUp(completion: @escaping (Error?) -> Void) {
-        
-        ArticleTestHelpers.setupWithNetworkFixtures {
-            completion(nil)
+        defer {
+            ArticleTestHelpers.tearDownNetworkFixtures()
         }
-        
-    }
 
-    override func tearDown() {
-        ArticleTestHelpers.tearDownNetworkFixtures()
-        super.tearDown()
-    }
-
-    func testArticleVCAccessesSchemeHandler() throws {
-        
         // test that articleVC converts articleURL to proper scheme and sets up SchemeHandler to ensure it is accessed during a load
-        
-        let tempDatabaseExpectation = expectation(description: "Waiting for temp database setup")
-        
-        var dataStore: MWKDataStore!
-        MWKDataStore.createTemporaryDataStore { result in
-            
-            dataStore = result
-            tempDatabaseExpectation.fulfill()
+        let dataStore = await withCheckedContinuation { (continuation: CheckedContinuation<MWKDataStore, Never>) in
+            MWKDataStore.createTemporaryDataStore { result in
+                continuation.resume(returning: result)
+            }
         }
-        
-        wait(for: [tempDatabaseExpectation], timeout: timeout)
-        
+
         let theme = Theme.light
-        let url = URL(string: "https://en.wikipedia.org/wiki/Dog")!
+        let url = try #require(URL(string: "https://en.wikipedia.org/wiki/Dog"))
         let schemeHandler = MockSchemeHandler(scheme: "app", session: dataStore.session)
-        guard let articleVC = ArticleViewController(articleURL: url, dataStore: dataStore, theme: theme, source: .undefined, schemeHandler: schemeHandler) else {
-            XCTFail("Failure initializing Article View Controller")
-            return
-        }
-        
-        let setupExpectation = expectation(description: "Waiting for article initial setup call")
-        
-        articleVC.initialSetupCompletion = {
-            setupExpectation.fulfill()
-            XCTAssert(schemeHandler.accessed, "SchemeHandler was not accessed during article load.")
+        let articleVC = try #require(ArticleViewController(articleURL: url, dataStore: dataStore, theme: theme, source: .undefined, schemeHandler: schemeHandler))
+        defer {
             UIApplication.shared.workaroundKeyWindow?.rootViewController = nil
             dataStore.clearTemporaryCache()
             dataStore.session.teardown()
         }
-            
-        UIApplication.shared.workaroundKeyWindow?.rootViewController = articleVC
-    
-        wait(for: [setupExpectation], timeout: timeout)
-    }
 
+        await withCheckedContinuation { (continuation: CheckedContinuation<Void, Never>) in
+            articleVC.initialSetupCompletion = {
+                continuation.resume()
+            }
+
+            UIApplication.shared.workaroundKeyWindow?.rootViewController = articleVC
+        }
+
+        #expect(schemeHandler.accessed, "SchemeHandler was not accessed during article load.")
+    }
 }
